refactor: reorganize files and update paths
