+++ src/Views/Pages/Manage.php
