+++ src/Views/Pages/TakeExercise.php
